refactor: move Mock to pkg/browser/browsertest subpackage
Mock and Call were compiled into every binary importing pkg/browser.
Move them to a dedicated browsertest subpackage (following the
net/http/httptest convention) so test helpers stay out of production
binaries. Add compile-time interface check in browsertest/mock.go.
